requirements.txt
openai
langchain
langchain-community
langchain-openai
faiss-cpu
streamlit
python-dotenv
requests
PyPDF2